Add basic assembly finishing feature (close #74)
TODOs left for this feature (filing under the purview of #72) --

-Make scaffolds play nice with this (since we set autounselectify
to true while in finishing mode). Maybe have scaffolds just highlight
their nodes a certain color (via style classes or something)?
Basically, I want scaffolds to be viewable during finishing mode.
Right now the problem is that they aren't.

-Add undo/redo buttons (for adding nodes), I guess?

-Optionally (?) enforce that paths can only be constructed from
adjacent nodes (e.g. if a node A is the last node on the path, then
node B can only be added to the path if there exists an edge from
A to B).