Remove DNA export button (close #126, close #114)
This feature was supplanted by the finishing feature, which exports
node IDs instead of actual DNA. If we want to add this back for
some reason, it shouldn't be too hard to do (after all, the
pertinent code is in the git history -- just check the diffs on this
commit to see what would need to be reverted).

Some related things I can do from here on -- #127, #145, etc.