Save basic SPQR tree data; output .gv views #173
Not quite sure about the trees being drawn (the MaryGold test, for
example, seems to be a bit different from the diagram, but that might
just be due to edge directionality not being a consistent thing in
SPQR trees? will look into it). This is a good start, though.

(The .gv representation is just a demonstration, by the way.
Eventually I'll get to the point of outputting this stuff to the .db
file and using that to visualize (and support collapsing/
uncollapsing) SPQR trees in the viewer application.)