Update Cytoscape.js to version 3.1.0
Included in this commit are changes to my code (and updates to
the dagre and cose-bilkent layouts) to make the web tool work with
the new version of Cytoscape.js.

I looked at the "upgrade guide" for v3 of Cytoscape.js, so I think
pretty much all of the issues detailed there should be addressed.

(Of course, there's certainly a chance I missed something)