Store metanode edge info, begin m.n. layout work
(#173) Still need to work on getting virtual edges oriented properly.

Once all internal edges are oriented correctly, the metanode internal
layouts will be accurate. At that point we'll be able to construct
Edge objects for the internal edges in the metanodes, and store their
relative control point information as attributes.

After that, the next action here will be to determine the orientation
of edges between metanodes.

Once that's done, I can use the information obtained in
layout_isolated() to lay out a compressed graph, in which every
biconnected component is replaced with a cluster which, in turn,
contains a tree of metanodes in which each metanode is just a
rectangle (same as with normal backfilling).

Then I can get the coordinates of each backfilled metanode and store
them, using these coordinates to set the absolute coordinates of the
internal nodes and edges of all metanodes.

At that point, the layout information for the SPQR view will be
complete on this end. I'd just have to display the compressed graph
(even more compressed -- i.e. only show root node of the SPQR tree
in the viewer at first, uncollapsing it on click) in the viewer.