Detect leaf metanodes (#173)
TODO -- just need to use this information to "propagate" changes
to the orientation of the edges up within the tree.

e.g. for a tree of

   S'
  / \     e2
  P R
 /\  \    e1
S" S  S"'

If we identify S", S, and S"' as the leaves, then we can classify
all edges on level e1 as pointing toward S", S, and S"' respectively.
Furthermore, we can use that information (combined with a few
simple axioms of trees) to orient the edges on level e2, also.
This process can be repeated until we have the root of the tree
correctly identified.