Highlight non-overlapping GML edges (#190)
Since it seems like the vast majority of edges in at least the
GML assembly graphs I'm working with now are overlaps, I figure using
the fancy option for non-overlapping edges is a more efficient
approach (and sort of makes sense if you think of non-overlapping
edges as "fragmented", I guess?).

Work remains to be done here to make this make sense for GFA and
LastGraph files. See #190 for more info.