Get "single graph" edges (#10, #173) and .gv test
We can eventually use these edges to draw "single graphs" using, say,
neato/circo/fdp/sfdp/twopi. (Of course, that'd be using PyGraphViz.)

Furthermore, single graph information like this will be useful in
generating SPQR trees (since the input graphs for those are assumed
to be undirected) for the corresponding graph.

Also it might be a cool idea to eventually draw single graphs with
"double" edges, where instead of using 2 nodes we just give
each edge a specific headport/tailport configuration. However, that's
not really a super important feature at present -- so not going to
bother with that for now. (something to think of re: #10)