Root SPQR trees at metanode with max |V| + |E|
This heuristic seems to work well for the test graphs I've tried
it out on thus far. Based on all I've read in the past few weeks,
it seems there isn't a panacea for this problem; so an approach
like this is probably fine for most cases.

(And, of course, it'd be really easy to modify this criteria in the
future if necessary.)

This commit (combined with the decision to generate SPQR trees from
single graphs -- will take care of that in the immediate future)
closes #186, closes #187, and closes #188. Next up is actually
passing the single graph info to the SPQR script, taking into account
various subtleties w/r/t filetypes, arguments, and files generated.
After that we can lay out the internal structure of metanodes,
perform backfilling, lay out the internal structure of the SPQR
trees for each bicomponent, perform backfilling, and then put all
that info in the .db.