Only create _single_links files for non-GML input
For #173.

For GML files, we can just pass both -s and -t to the SPQR script
(since contigs are already oriented in those graphs -- so the default
action we take is drawing a "single" graph -- since drawing a
"double" graph would just generate extra connected components).
Therefore, generating a _single_links file is redundant for a GML
file (since its contigs are already oriented).

For GFA and LastGraph files, though, we can't make that assumption.
Contigs in these graphs usually have explicitly given reverse
complement contigs, and it's possible to have both orientations of a
contig in the same connected component. (Looking at the GFA files
generated by Bambus 3, a defining feature is that every contig is
only explicitly used in a single orientation in the graph. This is
different from the assembly graphs generated by Velvet, since the
edges in those graphs can explicitly reference different orientations
of the same contig.) Therefore we need to generate a _single_links
file for constructing the SPQR tree for GFA/LastGraph files.

What's next to do is using the distinct_single_graph flag variable
to determine how to call the SPQR script. If it's False, then
we just need to call the script once with -s and -t, as detailed
above. If it's True, then we should call the script once with the
double graph edges (with -s) and once with the single graph edges
(with -t). Anyway, after that's taken care of we can actually
parse the SPQR tree stuff and associate it with the single graph.