Revert edge thickness scaling to global in c.comp
(See #169 for context.)

After a lot of playing around with the old way of doing this, I'm
going with the old way -- eventually we'll adjust it to avoid
problems resulting from outliers (see #184).

The rationale for this decision is that although the old method
helped with identifying paths, it didn't give an immediate visual
overview of edge weights in the current connected component:
One edge might look really thick, and another
might look really thin, but the thin one might actually have a
higher weight than the other edge due to the "thin" edge originating
from a node with a bunch of really high-weight outgoing edges or
something.

This scaling method ensures that any two edges in the currently
drawn connected component are visually comparable. Which is nice.

And, like I mentioned, once outlier detection and labelling is
implemented, this could be even more useful.

(Also we might want to re-enable this method in the future -- if
so it's like a line of code to change edge thicknesses, so that'd
be very doable)