Alter Frayed Rope starting prefix
Note that this change hasn't been pushed on the python side of
things yet since I'm still working on that.

The reason I'm doing this is to make working with SPQR metanode
IDs easier (since there'd be a conflict otherwise of R metanodes
vs. frayed Ropes). Now, there shouldn't be ambiguity (at least for
the foreseeable future -- and after a certain point we could switch
to not encoding this information in IDs, anyway).