Run DFS on single graph; compute its c.comps #173
Next up is actually laying out the single graph's connected
components. This time, the structures that will be backfilled are
Bicomponents.

After we do the layout, and get its information, and store that in
the .db then... I guess we're done on the python side of things?
After that, just need to interpret that layout in the JS side of the
tool. Should be ok.