Assign UUIDs to metanodes #209
This makes the IDs of metanodes containg a ton of children take up
a lot less space.

I think (?) it might have helped with #209? Shakya cmp 1 now loads
fairly quickly. But it's still not that great (see Biofilm 1, cmp 1
for an example of one where the extra compression doesn't seem to
be helping) -- there are probably other factors contributing that
I can make more efficient.