Resolve a cluster init bug in viewer; .db changes
Now, node_counts are stored in metanodes and bicomponents' entries
in the database.

Basically the tl;dr here is I was getting a really weird error from
Cytoscape.js' renderer that metanodes' dimensions were being set
as equal to negative infinity. Upon further inspection, it turns out
that not setting any children of metanodes, but still treating them
as nodes with children (i.e. in initClusters()) was causing this
problem (since the width of a metanode is calculated from the widths
of its children, my code wound up calling log(0) somewhere,
hence the negative infinity).

I resolved this bug by just not calling initClusters() (and,
moreover, just not adding any metanodes/bicmps to _uncollapsed)
when drawing SPQR stuff. For displaying cluster node count (which
I noticed was bugged), we use the values stored in the database.