quick optimization in collate.py
(dunno why I even wrote it the old way in the first place)
Basically, the old code checked if the set of nodes was a valid
source/sink pair, and then applied the "if |nodes| > 10 then
reject it and break from the loop" condition afterwards.

But that's a really silly way to go about that, since the length of
the collection of bubbles isn't going to change or anything while
inside an iteration of the loop -- so there's no reason not to just
apply the break condition before checking if the set of nodes
constitutes a valid source/sink pair, instead of afterwards.