Fix SPQR script bug in bicomponent tree creation
Turns out the error was that the tree_index variable was being
initialized and reset to 1 every connected component? So ensuring
that it was only initialized once (before iteration over connected
components started) seemed to fix this problem.

While attempting to fix this bug, I noticed & fixed a potential bug
in the way connected component "startNodes" are assigned, but
that's a separate issue for another day.