Darken edges and explicitly draw them above nodes
Kind of a subjective design choice (we can definitely change this
in the future if requested), but I think I like this better than
the previous drawings. It's a lot less ambiguous where edges
are routed (no more worrying about "is that 2 edges incident on
one node or one edge that passes through that node?").
The main downsides to this approach that I've identified thus far are

1) when there are lots of edge overlaps on nodes
the graph starts to look a bit ugly

2) there doesn't seem a way to explicitly set the z-index of node
labels without also changing the z-index of nodes, so now all edges
are drawn above node labels (I'd sort of prefer node labels to be
on a higher z-index). But this might be rectifiable by
messing with node label placement.