Center labels inside nodes
Labels are now also colored white when a node is GC-content colorized
or selected, so that the label shows up against the darker backdrop.
(Also, it looks cool to see the label change color.)

In a similar vein of stylistic enhancements, it might be cool to
highlight selected nodes by adding a border to them. (There's
commented-out code in xdot2cy.js that does this, actually.) However,
according to Cytoscape.js' docs, node borders can be expensive -- so
I think it might be more performant to just darken nodes rather than
drawing a border and everything. (Furthermore, the fancy label stuff
only applies when labels are drawn, which [thanks to the
min-zoomed-font-size parameter] is when they're pretty close-up, i.e.
when not a lot of nodes are on screen at once.)