Make collapsed cycl.chains repeatable: close #193
So I was getting this weird error where adding a cyclic chain would
add it an increasing number of times (once the first time I ran
finishing on a given connected component, twice the second time,
thrice the third time, etc.). Turns out I wasn't calling cy.off()
correctly, so the binding of

cy.on("tap", "node", addNodeFromEventToPath)

wasn't actually getting removed. This had the effect of creating
multiple redundant bindings of addNodeFromEventToPath(), meaning
that every time we'd tap a node addNodeFromEventToPath() would
be called n times (where n is the number of times finishing had
been started).

The reason I didn't see the effects of this problem until now
was because we have a check in addNodeFromEventToPath()
that ensures that the clicked node is actually a "tentative" node.
This error actually was visible in one-node cyclic chains previously,
but it seems I never bothered checking those to see if that was a
problem. (My fault.)

In any case, the problem's resolved now (finishing should be somewhat
more efficient now, since there aren't a bunch of redundant bindings)
and, now that collapsed cyclic chains are repeatable during
finishing, we can declare #193 as resolved.