Show GC stuff based on DNA_AVAILABILITY: fix #226
This also has the nice side effect of allowing GFA files with no
DNA sequences given to be used in MetagenomeScope. (That was doable
before, but it didn't work 100% correctly due to GC content stuff
making assumptions that all GFA files had DNA passed.)

We could also extend this functionality in the future for other DNA-
reliant statistics.