Add (disabled) code for bin GC-colorization #207
This is kind of neat to play with. There's an issue with the way
the upper part of the colorization is set (see the added comments
for a discussion of that). Also I guess this should eventually
have a UI option to switch between binned colorization and absolute
colorization, as well as to set the bin sizes.

In any case, something to think about.