Make colorization extreme colors modifiable #233
I'd still like to work on adding a small preview for the gradient
if possible (something like 0%, 25%, 50%, 75%, 100%?), so leaving
this open for now.

Also there seems to be a bug where the colorization previews stay
open if the modal dialog is closed with the Escape key -- I want to
fix that.