Close all colorpickers on dialog hide; close #233
nice.

there's ostensibly an inefficiency in that the binding uses the
general $(".colorpicker-component") selector -- meaning that
all colorpickers will be considered, not just the single open one
-- but I doubt this is a significant issue, since ~21 extra
iterations through a loop is not that much.

(That being said, if that inefficiency becomes a problem for some
reason, it should be fairly simple to also introduce bindings
on the "showPicker" colorpicker event that change a global variable
indicating the currently displayed colorpicker, which we can then
use to target the hide action on when the settings dialog is hidden.
But I don't think introducing that extra complexity is worth it
at present.)