Make only SPQR metanodes semitransparent
Changed the colors of the non-SPQR-metanode node groups (i.e.
structural patterns in standard mode, and bicomponent nodes in
SPQR mode) accordingly. This means that these node groups will match
their "normal" color when viewed on a white background, even when
the graph background is a different color.

The reason I preserved semitransparency in SPQR metanodes is to show
overlaps between metanodes overlaid on top of each other, which
happens fairly regularly in large bicomponents in implicit mode.