Small changes to settings/edge dialog text
To make the application's dialogs more consistent (re:
capitalization, etc).

There also wasn't really a need to keep the settings dialog named
"initialization settings", since all of the "settings" in the
application are contained in there at present (I'd classify the stuff
in the control panel as "controls", if that doesn't sound too
pedantic).