Support importing AGPs referencing node groups
Now, AGPs containing collapsed node groups exported from the viewer
interface can be imported as-is in the viewer interface.

This came with a slight cost in efficiency, since now all node group
IDs in standard mode have to be stored in COMPONENT_NODE_KEYS. But
that shouldn't make that big of a difference, and I haven't observed
any adverse impacts on drawing time thus far.