Add jQuery file to js/ folder: close #222 for now
Eventually we could look into supporting both a CDN and local file
for jQuery (and Bootstrap and d3.js, etc., I guess?) but I think just
having everything be local (as it is now) is a fair compromise that
will allow full offline use of the tool. Also, the minified jQuery
file is 92K (according to du -h), so I doubt that that's going to
be too heavy to download on most devices (especially compared to
sql.js, which is 2.1M).