Fail gracefully on invalid AGPs being loaded
NOTE that the only checks we really run on the validity of AGP
files are 1) that the extension is of type .AGP (case-insensitive)
and, now, 2) that each line in the file has at least 6 tab-separated
columns. We can add more checks in integrateAGPline() later if
desired, I suppose.

However, at least now (I think) we shouldn't get errors from invalid
AGP files being passed to the viewer -- before we try to operate
on the value of lineColumns[5], we check to make sure that that
result is defined (i.e. the line has >= 6 columns). If not, we raise
the offending line to the user.