Force duplicate file reload for .db & .tsv files
Closes #237!

I was already doing this with .agp files; not sure why I never
adapted it to .dbs.

Anyway, this basically forces the program to reload a file, even
if it's the same location/name/etc as the currently-loaded file.
This is useful because sometimes a user might want to make small
changes to an already-loaded file and then reload it; previously
(as I observed earlier with .tsv files), this didn't work. Now it
does.

(Also, clearing the .value attributes of file selectors saves some
memory (albeit a very tiny amount), so that's another upside to
this.)