Change node colorization inputs; clean that code
Now, we use inline radio buttons (c/o Bootstrap, of course) instead
of the button-styled radio buttons. This looks a lot nicer with
multiple elements, and the HTML for it is a lot easier to read.

Also, I cleaned up the changeNodeColorization() code to just
work directly with the classes instead of the complicated
define-a-function-for-each-colorization-class thing that I was
doing earlier.

Notable thing-that-should-be-changed-ASAP: right now, the "repeats"
colorization option is enabled for all GMLs. We should only enable
it for those graphs that have repeat detection enabled. (Also we
should actually get the is_repeat data for each node and use that
to generate the repeat colorization for each node, but that'll be
later tonight.)