fix minor error in node colorization radio btns
the "none" option wasn't getting checked when a new DB file was
loaded. it'd still eventually get checked when a component was drawn,
but there was a period where (vaguely, through the disabled-grayed-
out areas) you could see that the "none" option wasn't checked.

(Also, I cleaned up the code a bit -- turns out changing the
"checked" property of a radio button changes the "checked" properties
of the radio buttons in its same group, so you only need to change
the "checked" property of the "none" option instead of all the
options.)