Store is_repeat for all nodes in .db #165
For both nodes and singlenodes, so the colorization here should
work fine in the SPQR modes as well.

This also stores a field called "repeats_given" in the assembly
table, which functions similarly to "dna_given" for GC content.

The is_repeat value in the database can be either
0 (False), 1 (True), or NULL (the NULL is given if no annotation
was given for the node in question).

Anyway, all that's left for #165 is to make use of this data
in renderNodeObject()