Call resetColorSettings() before page unload
This prevents modified color settings from being preserved when
the page is reloaded via the forward/back browser buttons. Before
today that preservation thing wouldn't be a problem (since we
stored the default color settings as a string independent of the
DOM), but now we get the default color settings from the DOM's
state when the page is first loaded.

While that approach to getting the default color settings does
make updating the default color settings a lot easier, it assumes
that the state the page is initially loaded in has the correct
default settings applied. If modified colors are present when the
page is "loaded," then that'll cause those modified colors to be
assumed as the new default colors, which will mess up the
functionality of resetColorSettings().

Calling resetColorSettings() before the page is unloaded ensures
that when the page is reloaded, the original default colors are
used.