Temporarily don't give edges explicit IDs #258
The effect this has is ensuring that duplicate edges are rendered.

I guess the downside is addSelectedEdgeInfo() now has to do
four queries in std. mode (one for the source node then for its ID,
and one for the target node then for its ID) instead of one (for the
current edge ID, which the source ID/target ID were embedded in).

However, I went through the tool's use cases, and it looks like
nothing was broken by this. If I eventually decide to forbid
duplicate edges per #258 (and after this graph isn't needed any
more), then I can revert this behavior. For now, though, things
should be fine.