rm MAX_COMPONENTS & MIN_COMPONENT_SIZE configvars
Should probably tag the previously-closed issue #35 in this.
These weren't getting used anyway -- the SPQR layout processes
assumed that these weren't set.

If this functionality is requested, I can implement it
better (e.g. as a command-line option that actually has impact
on the SPQR processes). But I think not having it here is ok (and
the code with it present is in the history of the repository,
anyway).