Make edge hist bin ct. configurable for user #232
Other changes in this commit:

-Changed up the JS to not enable/disable the cullEdges stuff (since
 that will only be available if the filter edges stuff is available
 -- therefore, no need to worry about enabling/disabling sub-controls
 within the edge filtering dialog).
-Encapsulated the new bin ct. UI element and the cullEdges UI element
 in Bootstrap cols and a row, making the edge filtering dialog's
 formatting look a lot nicer than before.
-Moved the edge weight histogram drawing stuff into its own
 function to facilitate ease of redrawing the histogram.

Next up for #232 is improving the general UI of this (e.g. using
enter to redraw the histogram), maybe flipping the UI elements
so the buttons are on the right of the text inputs (would match
the search bar), making the chart look prettier (fixing occasional
bar spacing issues, removing floating-point tick values, making
sure that the max edge weight is still included in other bins (?),
etc.), and adding graphical filtering functionality.