Add MetagenomeScopeSite repo to docs/ folder
This has the effect of hosting the site on the GitHub pages page
for this repository.

If we're going to go ahead with this then I guess it might be
worth reconsidering the placement of some now-redundant files (e.g.
fonts/, Bootstrap source, etc). For the time being, though, this
is a cool setup.