Replace some invalid HTML (#79)
changes made include:
-Removed some of the <p> tags that contained <div>s -- it looks
 like a few of them didn't even have a visible impact on layout
-Replaced uses of the border attribute of <table>s with CSS --
 looks slightly different, but IMO it isn't a big deal and we can
 tweak the formatting in the future if desired
-Fixed the aria-controls attribute for tab links (in the settings
 dialog). Apparently it's supposed to point to the ID of the
 destination, not a description of the destination -- my fault,
 didn't read it carefully enough when I was implementing tabs in
 the settings dialog a few weeks back.

That should be mostly it for now. will continue to work on #79,
or at least the parts of it that I can fix (some of the
errors/warnings from the W3C validator are inevitable with Bootstrap,
which is fine -- but I figure those errors that I can fix, I probably
should).