Fix <p> nesting errors and colorpicker labels #79
The <p> nesting errors were due to <p> elements not being allowed
to contain <div>s in valid HTML. All of those errors should be
fixed now.

The colorpicker label errors were due to labels referring to the
colorpickers' <div>s instead of their inner <input>s (apparently
labels need to be associated with a form element, so referring to
a div in the form attribute is invalid). Thanks to some vim regex
magic, every colorpicker's inner input tag now has its own ID,
and each colorpicker's label refers to its respective ID now.
An added bonus of this -- clicking on a label for a given colorpicker
now focuses the browser on that colorpicker's text field, which
is actually pretty neat.

Might still try to fix a few things related to #79 (with the
exception of #80, all of the errors left seem due to my uses of
Bootstrap elements). will close this issue when done.