Use non-rounded, initial node dimensions: fix #78
This should make the code a bit easier to follow, also (in
retrospect, assigning heights/widths after layout was a bit silly).

Next up is #1. #78 was the main thing I needed to do before that,
so this shouldn't be too difficult to do now that I understand
Graphviz' rounding algorithms better.