Implement #1 for standard mode
Still need to implement it for the SPQR modes, although that
shouldn't be that bad to do.

Also I'm kinda waffling as to whether or not to preserve my fixes
for #78 -- the fact that Graphviz' edge control points and component
bounding boxes are both based on the rounded coordinates might mean
that it'd be easier (plus look nicer, plus be more consistent with
before) if I just performed the rounding method manually in the
faking process, and reverted the changes done in #78. I dunno. I'll
look into it.