Add functionality for user-supplied bubble files
Makes progress toward a resolution for #40 and #12. (I'd like to
do some more testing before closing either of these issues. Also,
for #12 to be resolved I'd like to add a generic "user-defined
pattern" NodeGroup class and use that in addition to user-defined
bubbles/etc. (get some feedback on what'd be useful and worth
spending time on)