Remove group _STYLE vars; start MiscPattern class
I wrote the _STYLE stuff back when the primary method of
visualizing files in MetagenomeScope was by examining dot's direct
output, instead of through the viewer interface. Now the _STYLE stuff
is essentially useless, so I removed it to make the code clearer.

I also started the skeleton of the MiscPattern class, which'll
eventually be able to store pattern type names, etc.