Identify user-supplied patterns in -up files
Still need to highlight these accordingly in the viewer interface.

Also, now all clusters in the .db file have a cluster_type attribute
describing their "type" (this is important for misc. patterns).

It might be a nice idea to treat certain cluster types (Bubbles,
Frayed Ropes, Chains, Cyclic Chains) specially, in that they'd just
get lumped with the other clusters of that type -- but for now just
putting all -up patterns as "misc" is ok. At this point that'd
probably be the best solution here, but I'd like to keep -ub around
to make MetaCarvel bubble exporting easier.