Move most of the README's docs to GitHub Wiki
This is already much, much more readable than before, and it'll
make updating this info a lot easier in the future.

(I kept the Acknowledgements/Contact/License sections of the wiki
in the README, since those are all pretty important and I don't
really think it'd be right to not provide that information with
the source code. Although that might be a silly reaction.)

Next up, I should also update the website to refer to the Wiki as
the canonical source of documentation for the tool.