(Temporarily?) comment out/rm SPQR code
Mostly from the python side of things -- a lot of stuff left in
the viewer interface JS.

Bubbles detected using SPQR trees in MetaCarvel can now be passed
in to MetagenomeScope via -ub, so those are covered.

Backups of the deleted/modified files in question (everything
affected by this commit) are still saved in the repository's
history, on my local system, and in the CBCB, so restoring this code
if desired should be easy.