Readd no-SPQR script versions as alternate option
Need to describe this a bit more, but this version can be used
portably on any system with Python/Graphviz/PyGraphviz installed
without the user needing to manually compile the C++ SPQR script,
install OGDF, etc. Until we (hopefully) get the SPQR stuff working
properly everywhere, this is a temporary measure to ensure
portability of the core standard-mode functionality of
MetagenomeScope.