Fix startNodes init error in SPQR script #86
Initialize all of startNodes to NULL manually instead of the fancy
way.

I still want to test this to make sure it works, but I don't have
access to a Mac system to test that this fixed the problem so I'm
not closing the issue just yet. Will downgrade its importance,
though, since this *should* resolve it.