Remove log files from repository root
Due to changes in the preprocessing script, these files have become
sort of irrelevant. I suppose the _with_optimizations log file is
useful in that it highlights the effects of #1 being resolved, but
IMO that's not really pertinent enough to justify it being included.

These files are still backed up on my system -- I guess I'll also
upload them to the CBCB computing infrastructure so that there's
multiple backups of them, in addition to the ones in the
repository history.