Support GFA files with interspersed links: fix #89
I also should get to work on adding a solution to #75, in the
vein of fully supporting the output of tools like fastg2gfa. But this
is a cool first-ish step in that direction!

This sort of solution might be desirable to implement for LastGraph
and MetaCarvel GML files, but since I don't think either of those
filestypes usually have links interspersed with node definitions
just supporting this in GFA files should be fine for now. (I can
make a separate issue for that later, if that's requested.)