Clean up MODAL_ACTIVE code: resolve #92
Now, all the dialogs' bindings for show/hide are all set together,
so future dialogs/etc. can just be added to the list of IDs in
doThingsWhenDOMReady().

Beyond that... I think this feature is taken care of. I'd like to
add even more keyboard support in the future (for example, panning
via keys?) but this is a good start.