Add -nt/--notriangulation option
So that users don't *have* to install GTS.

I'm running some tests now, though -- it looks like triangle smoothing
isn't super useful in a lot of cases, so it might be a better idea
to make no triangle smoothing the default and make the argument
*enable* triangulation (via changing it to -tri/--trianglesmoothing,
or something). Might implement that change later tonight/etc.