Fix minor bug in autofinishing tentative marking
In which, when a cycle was reached during autofinishing, tentative
nodes weren't marked as such (via the "tentative" class being added).

The problem is fixed now.

I also abstracted the code for marking tentative nodes to its own
function -- it's just 3 simple lines, and NEXT_NODES is a global
variable in xdot2cy.js so the function isn't very complicated.