Update minified files; add script to automate that
The commit.fish script can just be run in lieu of "git commit".
It automatically runs the minification script and adds the
corresponding files (since git add shouldn't [I think so, at least]
add files that haven't been changed, it's ok to just do this
for all the minified files all the time) if they've been changed.

Of course this'll probably break down if/when I get to a point
where I want to delay committing changes to the viewer interface
code for a bit (e.g. while I'm finishing something up on the
preprocessing script). In that case, reverting to just using
the normal git commit stuff should be fine.

(Also using this script might turn out to be really annoying. But
it's something worth trying, I guess!)