Remove padding in finishing animation fitting
A problem I noticed when using the finishing tools on a small-screen
device (e.g. smartphone) was that the animation for finishing seemed
to actually zoom out, instead of zooming in closer to the tentative
node(s). It turns out that the padding takes up more of the screen
on smaller devices, due to the padding being specified in pixels
(200px padding on a laptop screen is a lot more than 200px padding
on a smartphone screen).

So this brings up a problem: we want some padding around the
tentative node(s), if possible, but we also want the finishing
process to work well on arbitary-sized screens. I've gotta look into
ways of specifying padding in relative units, or maybe getting around
the problem in other ways (e.g. include some outgoing neighbors of
tentative nodes in the fitting operation if they're close by?).