Add functionality for replicating non-backfilling
via the -nbdf option. gotta talk about that in the wiki.

This is really cool, and will make writing about this in the paper
a lot easier. To produce DOT files that use backfilling -- for
comparison -- the -pg option works fine (it doesn't include
the child nodes inside cluster rectangles, but it should be
apparent that the rectangles are sized appropriately).