Change selected node color instead of darkening
This has the effect of 1) making the color change work well in
arbitrary colorschemes (or at least giving the user the options
to change it to make it work well) and 2) making node/edge selection
consistently styled.

The reason I originally used darkening (the background-blacken
node attribute in Cytoscape.js) was to preserve node colorization,
but now that all the colors in the graph display are configurable
that's kind of silly since the user can alter the colorization
gradient in a way that makes node colorization dependent on
darkness.