Update default colorscheme node lbls & edge z-idx
The labels' colors are being updated because, due to the new node
scaling methods, there are now a lot of small nodes in GML files
with long labels that run onto the graph background. Which is why
the new default node labels are gray instead of white.

Re: the edge z-index, that also has to do with node labels. And I
think it's probably the right choice to make (setting the edges
with a higher z-index than nodes used to mess with some things in
the graph), but I think it might be a good idea to make it
configurable?