Implement manual finishing max zoom animation #110
This is not an ideal solution. I think I've got another solution
ready that fixes this without the extra zoom-out animation, but I
wanted to get this committed so I have this code for reference
in the future (in case I want to do multiple animations in this
same sorta way, because I have to admit it looks kinda cool).