Make cluster kbd nav (#39) off by default
I think this is actually a better solution than enabling it by
default, because there's a high chance for users to press the
<-/->/A/D buttons without realizing what they're doing. And if
that happens, this can frustrate users since it's a relatively
sudden transition, and can move the user away from what they were
doing. So this feature should be enabled only when it's explicitly
requested.