Disable demo btn & enable if protocol ok; fix #109
Just having doThingsWhenDOMReady() call enableButton() on the demo
button resulted in a sort of "flash" effect, since the button would
go from its disabled status to its enabled status. So I implemented
a workaround where it starts at opacity 0 and its opacity is adjusted
via a gentler transition, which should look nicer (and avoid the
harsh-looking flash).

While I was typing the above commit message, I just realized that
I could set the button's opacity to 1 and adjust it down to 0.65 if
the protocol doesn't support CORS (or vice versa). There's no need
to start at 0... and that'd look nicer.

There seems to be a lot of information online on addressing this sort
of phenomenon (called the "Flash of unstyled content," or FOUC for
short). Even with the fix to this mentioned in the above paragraph,
I'm sure there are better ways to address this. I'll look into it.
Might set up another GitHub issue for this.