Add tabindex="-1" to file select dialog
To make it consistent with the other dialogs in the viewer interface.
Basically, this sort of forces the modal to take precedence in focus.
This results in some interactions with the dialog being immediately
supported when it's first shown, instead of only after it's
explicitly focused on by the user. One example (where I first
realized something was off) is pressing the Escape key to exit the
modal.