Update on demo button: start at disabled opacity
As is discussed in the message for commit
fa9e8987ce8f5b7fd8bbbc7186a5e2b64d3a4aac.

It might be desirable to flip this around, since I imagine the
viewer interface will probably be used with the http/https protocols
(i.e. hosted on a server somewhere) more than the other way around.
But that's kinda splitting hairs -- not super important.

(If we do go forward with that, I added a comment in xdot2cy.js that
discusses that approach and how to negotiate its issues a bit.)