Wait I forgot to add the actual commented files
My b (I was using that commit.fish script, which only adds the
minified versions).

The reason the minified JS file differed is because this commit
(well it would've been the last commit) removes the line that
sets the demo .db button's opacity to 1 if demos are supported.
This is because enabling the button already does this!

This is actually really nice, because it allows us to remain agnostic
to whatever the opacity levels set to enabled/disabled buttons by
Bootstrap are (well, assuming that they're at least different).
The way CSS transitions work, any change to opacity will result in
the transition animation occurring.