Make transitions only apply to demo btn 1st time
So that when all buttons are enabled/disabled together, they'll
all change together (instead of the demo button fading gradually,
and the other buttons changing immediately).

Might be worth making all buttons have the fading animation. Want
to think about it a bit more; I think #70 would be a good place
to look into that for?